user-story-tests/26-page-styles-only-affecting-shadow-tree.html

<!DOCTYPE html>
<html lang="en">

<head>
    <meta charset="UTF-8">
    <meta name="viewport" content="width=device-width, initial-scale=1.0">
    <title>Page styles only affecting shadow tree</title>
    <script src='./resources/testharness.js'></script>
    <script src='./resources/testharnessreport.js'></script>
</head>

<body>

    <style>
        @layer button-group-styles {
            :host(button-group) button {
                border: thick dashed red;
            }
        }
    </style>
    <button>Button outside a shadow tree (not thick dashed red)</button>
    <button-group>
        <template shadowrootmode="open">
            <style>
                @layer inherit.button-group-styles.as.page-defined-styles, page-defined-styles;
            </style>
            <button>
                Button inside a shadow tree (thick dashed red) (from page layer
                'button-group-styles')
            </button>
        </template>
    </button-group>


    <script type="module">
        import ShadowLayers from '../ShadowLayers.js';

        window.onload = function () {

            const elName = "button-group"

            test(function () {

                const inheritResult = ShadowLayers.inheritFromShadowStatementRule(document.querySelector(elName))
                assert_true(inheritResult, "inherit found, so returns true");
            }, "Page styles only affecting shadow tree");

        }
    </script>

    <div id="log"></div>

</body>

</html>
